Implement ball reset
Ball reset script is attached to the ball object. Upon hitting a goal, the goal script passes the ball it collided with into a GameManager method.
The game pauses for 2 seconds with text display announcing a goal, then resets the ball's position (a reset point  game object is attached to the ball's reset script) and velocity.